+++ src/main/resources/db/migration/V1_3__create_demands_table.sql
@@ -2,9 +2,9 @@ CREATE TABLE IF NOT EXISTS demands
 (
   demand_id INT NOT NULL AUTO_INCREMENT,
   consumer_id  INT NOT NULL,
-  amount      FLOAT  NOT NULL,
-  covered     FLOAT  NOT NULL,
-  price float not null ,
+  amount      DOUBLE  NOT NULL,
+  covered     DOUBLE  NOT NULL,
+  price DOUBLE not null ,
 PRIMARY KEY (demand_id),
 FOREIGN KEY(consumer_id)REFERENCES consumers(id)
 );
